% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goadDifGenes.R
\name{goadDifGenes}
\alias{goadDifGenes}
\title{Get differentially expressed genes in GOAD datasets}
\usage{
goadDifGenes(ids, p, foldChange = c("ALL", "UP", "DOWN"))
}
\arguments{
\item{ids}{GOAD ids of the datasets. To see which id refers to which dataset see GOADdatasets table}

\item{p}{p value threshold for selection. A threshold of 1 can be used to axtract all genes in the platforms}

\item{foldChange}{direction of fold change.}
}
\description{
Acquires differentially expressed genes in chosen GOAD datasets. Datasets are represented 
as ids in the GOAD database. To see which id refers to which dataset see GOADdatasets table
}

